Correct the types of built in scalars in querygen.

+++ cynic-querygen/src/type_ext.rs
@@ -30,8 +30,8 @@ fn type_spec_imp<'a>(ty: &Type<'a, &'a str>, nullable: bool) -> Cow<'a, str> {
     }
 
     match ty {
-        Type::NamedType("Int") => Cow::Borrowed("i32"),
-        Type::NamedType("Float") => Cow::Borrowed("f32"),
+        Type::NamedType("Int") => Cow::Borrowed("i64"),
+        Type::NamedType("Float") => Cow::Borrowed("f64"),
         Type::NamedType("Boolean") => Cow::Borrowed("bool"),
         Type::NamedType(s) => Cow::Borrowed(s),
         Type::ListType(inner) => Cow::Owned(format!("Vec<{}>", type_spec_imp(inner, true))),
